@@ -18,8 +18,8 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Energinet.DataHub.WebApi.Tests.Fixtures;
-using Energinet.DataHub.Wholesale.Application.Batches;
 using Energinet.DataHub.Wholesale.Client;
+using Energinet.DataHub.Wholesale.Contracts;
 using FluentAssertions;
 using Moq;
 using Xunit;
