@@ -17,8 +17,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Net.Mime;
 using System.Threading.Tasks;
-using Energinet.DataHub.Wholesale.Application.Batches;
 using Energinet.DataHub.Wholesale.Client;
+using Energinet.DataHub.Wholesale.Contracts;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Energinet.DataHub.WebApi.Controllers
@@ -63,6 +63,31 @@ namespace Energinet.DataHub.WebApi.Controllers
         {
             var stream = await _client.GetZippedBasisDataStreamAsync(batchId);
             return File(stream, MediaTypeNames.Application.Zip);
+        }
+
+        /// <summary>
+        /// Get a batch.
+        /// </summary>
+        [HttpGet("Batch")]
+        public async Task<ActionResult<BatchDtoV2?>> GetBatchAsync(Guid batchId)
+        {
+            var batchDto = await _client.GetBatchAsync(batchId);
+            return batchDto;
+        }
+
+        /// <summary>
+        /// Get a processStepResult.
+        /// </summary>
+        [HttpGet("ProcessStepResult")]
+        public async Task<ActionResult> GetAsync(Guid batchId, string gridAreaCode, ProcessStepType type)
+        {
+            var processStepResult = new ProcessStepResultDto(
+                MeteringPointType.Production,
+                new decimal(1.1),
+                new decimal(1.1),
+                new decimal(1.1),
+                new TimeSeriesPointDto[] { new TimeSeriesPointDto(new DateTimeOffset(DateTime.Now), new decimal(1.1)) });
+            return await Task.FromResult<ActionResult>(Ok(processStepResult));
         }
     }
 }
