@@ -16,6 +16,7 @@ using System;
 using System.IO;
 using System.Reflection;
 using System.Text.Json.Serialization;
+using Energinet.DataHub.Core.App.WebApp.Authentication;
 using Energinet.DataHub.Core.App.WebApp.Diagnostics.HealthChecks;
 using Energinet.DataHub.Core.App.WebApp.Middleware;
 using Energinet.DataHub.WebApi.Registration;
@@ -89,9 +90,10 @@ namespace Energinet.DataHub.WebApi
                 config.AddSecurityRequirement(securityRequirement);
             });
 
-            var openIdUrl = Configuration.GetValue<string>("FRONTEND_OPEN_ID_URL") ?? string.Empty;
-            var audience = Configuration.GetValue<string>("FRONTEND_SERVICE_APP_ID") ?? string.Empty;
-            services.AddJwtTokenSecurity(openIdUrl, audience);
+            var externalOpenIdUrl = Configuration.GetValue<string>("FRONTEND_OPEN_ID_URL") ?? string.Empty;
+            var internalOpenIdUrl = Configuration.GetValue<string>("INTERNAL_OPEN_ID_URL") ?? string.Empty;
+            var backendAppId = Configuration.GetValue<string>("FRONTEND_SERVICE_APP_ID") ?? string.Empty;
+            services.AddJwtBearerAuthentication(externalOpenIdUrl, internalOpenIdUrl, backendAppId);
 
             if (Environment.IsDevelopment())
             {
@@ -136,7 +138,7 @@ namespace Energinet.DataHub.WebApi
 
             if (!Environment.IsDevelopment())
             {
-                app.UseMiddleware<JwtTokenMiddleware>();
+                app.UseAuthentication();
             }
 
             app.UseAuthorization();
