@@ -22,7 +22,10 @@ import {
   ChargesHttp,
 } from '@energinet-datahub/dh/shared/domain';
 import { Observable, switchMap, tap } from 'rxjs';
-import { ErrorState, LoadingState } from './states';
+import {
+  ErrorState,
+  LoadingState,
+} from '@energinet-datahub/dh/shared/data-access-api';
 
 interface MarketParticipantState {
   readonly marketParticipants?: Array<MarketParticipantV1Dto>;
