@@ -42,6 +42,7 @@ import {
 import { WattCheckboxModule } from '@energinet-datahub/watt/checkbox';
 import { WattBadgeModule } from '@energinet-datahub/watt/badge';
 import { WattButtonModule } from '@energinet-datahub/watt/button';
+import { WattTopBarComponent } from '@energinet-datahub/watt/top-bar';
 
 import { PushModule } from '@rx-angular/template';
 import { DhChargesPricesResultScam } from './search-result/dh-charges-prices-result.component';
@@ -157,18 +158,19 @@ export class DhChargesPricesComponent implements OnInit, OnDestroy {
   declarations: [DhChargesPricesComponent],
   imports: [
     CommonModule,
+    DhChargesPricesResultScam,
+    FormsModule,
+    PushModule,
+    TranslocoModule,
+    WattBadgeModule,
     WattButtonModule,
+    WattCheckboxModule,
+    WattDatepickerModule,
+    WattDropdownModule,
     WattFormFieldModule,
     WattInputModule,
-    WattCheckboxModule,
-    WattBadgeModule,
-    WattDropdownModule,
     WattSpinnerModule,
-    TranslocoModule,
-    FormsModule,
-    DhChargesPricesResultScam,
-    WattDatepickerModule,
-    PushModule,
+    WattTopBarComponent,
   ],
 })
 export class DhChargesPricesScam {}
