@@ -15,14 +15,13 @@
  * limitations under the License.
  */
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { render, within } from '@testing-library/angular';
+import { render } from '@testing-library/angular';
 
 import { en as enTranslations } from '@energinet-datahub/dh/globalization/assets-localization';
 import { getTranslocoTestingModule } from '@energinet-datahub/dh/shared/test-util-i18n';
 import {
   getByTerm,
   getDefinitonByTerm,
-  queryByTerm,
   runOnPushChangeDetection,
 } from '@energinet-datahub/dh/shared/test-util-metering-point';
 import { MeteringPointType } from '@energinet-datahub/dh/shared/domain';
@@ -69,96 +68,6 @@ describe(DhMeteringPointPrimaryMasterDataComponent.name, () => {
 
     await runOnPushChangeDetection(fixture);
   }
-
-  describe('Address', () => {
-    it('should render address term', async () => {
-      await setup();
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.address
-      );
-      expect(term).toBeInTheDocument();
-    });
-
-    it('should show address with floor and suite number correctly', async () => {
-      await setup();
-
-      const term = queryByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.address
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(definition).toBeInTheDocument();
-      expect(definition?.innerHTML).toEqual(
-        '[streetName] [buildingNumber], [floorIdentification]. [suiteNumber]<br>[citySubDivisionName]<br>[postalCode] [cityName]'
-      );
-    });
-
-    it('should show address without floor and suite number correctly', async () => {
-      await setup({
-        ...testData,
-        floorIdentification: undefined,
-        suiteNumber: undefined,
-      });
-
-      const term = queryByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.address
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(definition).toBeInTheDocument();
-      expect(definition?.innerHTML).toEqual(
-        '[streetName] [buildingNumber]<br>[citySubDivisionName]<br>[postalCode] [cityName]'
-      );
-    });
-  });
-
-  describe('Is actual address', () => {
-    it('should render success icon, when actual address', async () => {
-      await setup();
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.actualAddress
-      );
-      expect(term).toBeInTheDocument();
-      within(term).getByRole('img', { name: 'success' });
-    });
-
-    it('should render actual address definition', async () => {
-      await setup();
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.actualAddress
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(
-        enTranslations.meteringPoint.overview.primaryMasterData.actualAddress
-      );
-    });
-
-    it('should render warning icon, when not actual address', async () => {
-      await setup({ ...testData, isActualAddress: false });
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.notActualAddress
-      );
-      expect(term).toBeInTheDocument();
-      within(term).getByRole('img', { name: 'warning' });
-    });
-
-    it('should render not actual address', async () => {
-      await setup({ ...testData, isActualAddress: false });
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.notActualAddress
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(
-        enTranslations.meteringPoint.overview.primaryMasterData.notActualAddress
-      );
-    });
-  });
 
   describe('Location description', () => {
     it('should render location description term', async () => {
@@ -265,119 +174,6 @@ describe(DhMeteringPointPrimaryMasterDataComponent.name, () => {
       );
       const definition = getDefinitonByTerm(term);
 
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(fallbackValue);
-    });
-  });
-
-  describe('Has electricity supplier', () => {
-    it('should render term', async () => {
-      await setup();
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData
-          .hasElectricitySupplier
-      );
-      expect(term).toBeInTheDocument();
-    });
-
-    it('should render definition', async () => {
-      await setup();
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData
-          .hasElectricitySupplier
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(enTranslations.yes);
-
-      const since = definition?.nextElementSibling;
-      expect(since).toHaveTextContent(
-        `${enTranslations.meteringPoint.overview.primaryMasterData.since.replace(
-          '{{date}}',
-          '17-12-2021'
-        )}`
-      );
-    });
-
-    it('should render fallback definition, if supply start is undefined', async () => {
-      await setup({ ...testData, supplyStart: undefined });
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData
-          .hasElectricitySupplier
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(enTranslations.no);
-
-      const since = definition?.nextElementSibling;
-      expect(since).toHaveTextContent(fallbackValue);
-    });
-
-    it('should render fallback definition, if supply start is null', async () => {
-      await setup({ ...testData, supplyStart: null });
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData
-          .hasElectricitySupplier
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(enTranslations.no);
-
-      const since = definition?.nextElementSibling;
-      expect(since).toHaveTextContent(fallbackValue);
-    });
-  });
-
-  describe('Meter ID', () => {
-    it('should render term', async () => {
-      await setup();
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.meterNumber
-      );
-      expect(term).toBeInTheDocument();
-    });
-
-    it('should render definition', async () => {
-      await setup();
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.meterNumber
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(term).toBeInTheDocument();
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(testData.meterId as string);
-    });
-
-    it('should render fallback definition, if meter ID is null', async () => {
-      await setup({ ...testData, meterId: null });
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.meterNumber
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(term).toBeInTheDocument();
-      expect(definition).toBeInTheDocument();
-      expect(definition).toHaveTextContent(fallbackValue);
-    });
-
-    it('should render fallback definition, if meter ID is empty string', async () => {
-      await setup({ ...testData, meterId: ' ' });
-
-      const term = getByTerm(
-        enTranslations.meteringPoint.overview.primaryMasterData.meterNumber
-      );
-      const definition = getDefinitonByTerm(term);
-
-      expect(term).toBeInTheDocument();
       expect(definition).toBeInTheDocument();
       expect(definition).toHaveTextContent(fallbackValue);
     });
