@@ -19,23 +19,12 @@ import {
   Component,
   Input,
   NgModule,
-  OnChanges,
-  SecurityContext,
-  SimpleChanges,
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { DomSanitizer } from '@angular/platform-browser';
 import { TranslocoModule } from '@ngneat/transloco';
 
 import { MeteringPointCimDto } from '@energinet-datahub/dh/shared/domain';
-import { DhSharedUiDateTimeModule } from '@energinet-datahub/dh/shared/ui-date-time';
-import { WattIcon, WattIconModule } from '@energinet-datahub/watt/icon';
-import {
-  DhEmDashFallbackPipeScam,
-  DhIsParentPipeScam,
-  DhShowForMeteringPointTypeDirectiveScam,
-  DhYesNoPipeScam,
-} from '@energinet-datahub/dh/metering-point/shared/ui-util';
+import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/metering-point/shared/ui-util';
 
 export type PrimaryMasterData = Pick<
   MeteringPointCimDto,
@@ -60,75 +49,13 @@ export type PrimaryMasterData = Pick<
   styleUrls: ['./dh-metering-point-primary-master-data.component.scss'],
   templateUrl: './dh-metering-point-primary-master-data.component.html',
 })
-export class DhMeteringPointPrimaryMasterDataComponent implements OnChanges {
+export class DhMeteringPointPrimaryMasterDataComponent {
   @Input() primaryMasterData?: PrimaryMasterData;
-  address?: string;
-  isActualAddressIcon: WattIcon = 'success';
-  actualAddressTranslationKey = 'actualAddress';
-
-  get electricitySupplierSinceTranslationKey(): string | undefined {
-    if (this.primaryMasterData?.supplyStart == null) {
-      return;
-    }
-
-    return 'meteringPoint.overview.primaryMasterData.since';
-  }
-
-  get isSupplyStartSet(): boolean {
-    return this.primaryMasterData?.supplyStart != null;
-  }
-
-  constructor(private domSanitizer: DomSanitizer) {}
-
-  ngOnChanges(changes: SimpleChanges) {
-    if (changes.primaryMasterData && changes.primaryMasterData.currentValue) {
-      const currentValue = changes.primaryMasterData.currentValue;
-
-      this.address = this.formatAddress(currentValue);
-      if (currentValue.isActualAddress) {
-        this.isActualAddressIcon = 'success';
-        this.actualAddressTranslationKey = 'actualAddress';
-      } else {
-        this.isActualAddressIcon = 'warning';
-        this.actualAddressTranslationKey = 'notActualAddress';
-      }
-    }
-  }
-
-  private formatAddress(data: PrimaryMasterData): string {
-    let address = `${data.streetName}`;
-    if (data.buildingNumber) {
-      address += ` ${data.buildingNumber}`;
-    }
-    if (data.floorIdentification || data.suiteNumber) {
-      address += ',';
-    }
-    if (data.floorIdentification) {
-      address += ` ${data.floorIdentification}.`;
-    }
-    if (data.suiteNumber) {
-      address += ` ${data.suiteNumber}`;
-    }
-    if (data.citySubDivisionName) {
-      address += `<br />${data.citySubDivisionName}`;
-    }
-    address += `<br />${data.postalCode} ${data.cityName}`;
-    return this.domSanitizer.sanitize(SecurityContext.HTML, address) as string;
-  }
 }
 
 @NgModule({
   declarations: [DhMeteringPointPrimaryMasterDataComponent],
-  imports: [
-    CommonModule,
-    WattIconModule,
-    TranslocoModule,
-    DhEmDashFallbackPipeScam,
-    DhSharedUiDateTimeModule,
-    DhYesNoPipeScam,
-    DhIsParentPipeScam,
-    DhShowForMeteringPointTypeDirectiveScam,
-  ],
+  imports: [CommonModule, TranslocoModule, DhEmDashFallbackPipeScam],
   exports: [DhMeteringPointPrimaryMasterDataComponent],
 })
 export class DhMeteringPointPrimaryMasterDataScam {}
