@@ -38,12 +38,12 @@ function getByGsrn(apiBase: string) {
   });
 }
 
-const getByGsrnResponse: MeteringPointCimDto = {
+export const getByGsrnResponse: MeteringPointCimDto = {
   meteringPointId: '10e01f8d-be5f-4253-80a2-37082baa950b',
   gsrnNumber: '573876630161457173',
-  streetName: 'Vestergade!!',
-  postalCode: '5500',
-  cityName: 'Middelfart',
+  streetName: 'Nørre Lysgade',
+  postalCode: '4545',
+  cityName: 'Fiktivby',
   countryCode: 'DK',
   connectionState: 'D03',
   meteringMethod: 'D02',
@@ -54,18 +54,18 @@ const getByGsrnResponse: MeteringPointCimDto = {
   gridAreaName: 'Volt U-001',
   gridAreaCode: '115',
   linkedExtendedMasterdata: '575774240560385464',
-  locationDescription: 'String',
+  locationDescription: 'Måler sidder bagved, under trappen',
   productId: 'EnergyActive',
   unit: 'KWH',
   effectiveDate: '2021-09-25T22:00:00Z',
-  meterId: '',
+  meterId: '123456',
   streetCode: '0405',
-  citySubDivisionName: 'Strib',
-  floorIdentification: '1',
+  citySubDivisionName: '',
+  floorIdentification: '4. tv',
   suiteNumber: '12',
-  buildingNumber: '10',
+  buildingNumber: '112N',
   municipalityCode: 625,
-  isActualAddress: true,
+  isActualAddress: false,
   darReference: '0a3f50b9-b942-32b8-e044-0003ba298018',
   capacity: 6000,
   assetType: 'D12',
