@@ -8,10 +8,14 @@ export * from './market-participant-grid-area-http.service';
 import { MarketParticipantGridAreaHttp } from './market-participant-grid-area-http.service';
 export * from './market-participant-grid-area-overview-http.service';
 import { MarketParticipantGridAreaOverviewHttp } from './market-participant-grid-area-overview-http.service';
+export * from './market-participant-user-http.service';
+import { MarketParticipantUserHttp } from './market-participant-user-http.service';
 export * from './message-archive-http.service';
 import { MessageArchiveHttp } from './message-archive-http.service';
 export * from './metering-point-http.service';
 import { MeteringPointHttp } from './metering-point-http.service';
+export * from './token-http.service';
+import { TokenHttp } from './token-http.service';
 export * from './wholesale-batch-http.service';
 import { WholesaleBatchHttp } from './wholesale-batch-http.service';
-export const APIS = [ChargeLinksHttp, ChargesHttp, MarketParticipantHttp, MarketParticipantGridAreaHttp, MarketParticipantGridAreaOverviewHttp, MessageArchiveHttp, MeteringPointHttp, WholesaleBatchHttp];
+export const APIS = [ChargeLinksHttp, ChargesHttp, MarketParticipantHttp, MarketParticipantGridAreaHttp, MarketParticipantGridAreaOverviewHttp, MarketParticipantUserHttp, MessageArchiveHttp, MeteringPointHttp, TokenHttp, WholesaleBatchHttp];
