@@ -8,7 +8,6 @@ export * from './actor-status';
 export * from './address-dto';
 export * from './asset-type';
 export * from './batch-dto-v2';
-export * from './batch-grid-area-dto';
 export * from './batch-request-dto';
 export * from './batch-search-dto';
 export * from './batch-state';
@@ -33,9 +32,13 @@ export * from './connection-type';
 export * from './contact-category';
 export * from './create-actor-contact-dto';
 export * from './create-actor-dto';
+export * from './create-charge-v1-dto';
 export * from './create-organization-dto';
 export * from './disconnection-type';
 export * from './eic-function';
+export * from './get-associated-user-actors-response-dto';
+export * from './get-token-request-dto';
+export * from './get-token-response-dto';
 export * from './grid-area-audit-log-entry-dto';
 export * from './grid-area-audit-log-entry-field';
 export * from './grid-area-dto';
@@ -58,6 +61,8 @@ export * from './process';
 export * from './process-detail';
 export * from './process-detail-error';
 export * from './process-status';
+export * from './process-step-type';
+export * from './process-type';
 export * from './product-id';
 export * from './reading-occurrence';
 export * from './resolution';
@@ -65,4 +70,3 @@ export * from './settlement-method';
 export * from './stream';
 export * from './unit';
 export * from './vat-classification';
-export * from './wholesale-process-type';
